requirements.txt
pillow==8.2.0
numpy==1.21.0
opencv-python==4.5.2.54
scipy==1.7.0
